requirements.txt
anyascii==0.3.1
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.1
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
decorator==5.1.1
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-filter==21.1
django-modelcluster==5.3
django-sass==1.1.0
django-taggit==2.1.0
django-treebeard==4.5.1
django==4.0.4
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
executing==0.8.3
gunicorn==20.1.0
html5lib==1.1
idna==3.3; python_version >= '3'
importlib-metadata==4.11.4; python_version < '3.10'
ipython==8.3.0
jedi==0.18.1
l18n==2021.3
libsass==0.21.0
markdown==3.3.7
matplotlib-inline==0.1.3
openpyxl==3.0.9
packaging==21.3
parso==0.8.3
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.12.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1; python_version >= '3.0'
sqlparse==0.4.2
stack-data==0.2.0
tablib==3.2.1
telepath==0.2
traitlets==5.2.0
urllib3==1.26.9
wagtail-markdown==0.8.0
wagtail==2.16.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.1.0
willow==1.4.1
xlrd==2.0.1
xlsxwriter==3.0.3
xlwt==1.3.0
zipp==3.8.0
